Main functions and sample Calspan data to test fitting
The raw data of few sample runs of Round 6 for cornering tests carried out by Calspan are provided, along with the necessary data clean-up and fitting routine codes. Use this to fit a data file to the MF5.2 tire model, then tune the vehicle dynamics parameters of your racecar and plot the Yawing moment diagram for it.